+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add new `RecurringSilence` CRD to create silences based on a cron expression.
+
 ## [0.17.0] - 2025-07-02
 
 ### Added